Added main player bullets and collisions
- Improved scene templated methods
- Added game config parameters to config bullets
- Bullets optim (Bullets have life time and are destroyed after the defined lifetime). 
- Fixed bug in collision module where collisions were not deleted when and entity was unspawned
- Reordered modules to avoid reading access violations given the ModuleCollision changes

+++ SDLGameTest/Engine.h
@@ -2,6 +2,7 @@
 
 #include<list>
 #include "Utils.h"
+#include "SDL.h"
 
 class Module;
 class ModuleRender;
@@ -48,6 +49,6 @@ private:
     static Engine* s_Instance;
     std::list<Module*> m_Modules;
     double m_DT = 0;
-    unsigned int m_Now = 0;
-    unsigned int m_Last = 0;
+    Uint64 m_Now = 0;
+    Uint64 m_Last = 0;
 };
